+++ src/common/route/ApiError.tsx
@@ -1,29 +1,32 @@
 import React from "react";
 import { isObject } from "lodash-es";
 import ReactJson from "react-json-view";
-import { AxiosError } from "axios";
 
 import Headers from "./Headers";
-import { useGlobalContext } from "../context/GlobalContext";
+import { useGlobalContext } from "../../context/GlobalContext";
 
-import Helpers from "../lib/helpers";
+import Helpers from "../../lib/helpers";
 
-export default function ApiError({ error }: { error: null | AxiosError }) {
+import { ApiError } from "../../types";
+
+export default function ApiError({ error }: { error: null | ApiError }) {
   const { darkMode } = useGlobalContext();
   if (!error?.response) {
     return null;
   }
 
   let responseColor;
-  if (error?.response?.status >= 200 && error?.response?.status < 300) {
+  if (
+    error?.response?.status &&
+    error.response.status >= 200 &&
+    error.response.status < 300
+  ) {
     responseColor = Helpers.colors.green;
-  } else if (error?.response?.status >= 400) {
+  } else if (error?.response?.status && error.response.status >= 400) {
     responseColor = Helpers.colors.red;
   }
 
   const isJSON = error.response.data && isObject(error.response.data);
-
-  const hasConfigHeaders = !!Object.keys(error?.config?.headers || {}).length;
 
   return (
     <div>
@@ -56,8 +59,6 @@ export default function ApiError({ error }: { error: null | AxiosError }) {
           {error.response.data}
         </div>
       )}
-
-      {hasConfigHeaders ? <Headers headers={error.config.headers} /> : null}
     </div>
   );
 }