+++ src/index.ts
@@ -1,13 +1,8 @@
 export { useGlobalContext } from "./context/GlobalContext";
-export { default as Route } from "./Route";
-export { default as Params } from "./route/Params";
 
 export { BadMagic } from "./BadMagic";
 
-export { default as BodyPreview } from "./route/BodyPreview";
-export { default as ApiResponse } from "./route/ApiResponse";
-export { default as ApiError } from "./route/ApiError";
-
+// UI elements that can be used with AuthForm
 export { default as TextInput } from "./common/TextInput";
 export { default as Select } from "./common/Select";
 export { default as Button } from "./common/Button";
@@ -16,8 +11,7 @@ export { default as Success } from "./common/Success";
 export { default as Error } from "./common/Error";
 export { default as Required } from "./common/Required";
 
+// Helper functions that can be used with interceptors
 export { default as Helpers } from "./lib/helpers";
-export { default as OpenApi } from "./lib/openapi";
-export { default as Storage } from "./lib/storage";
 
-export { Method, ParamType, Workspace as WorkspaceProps } from "./types";
+export { Method } from "./types";