+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 - Removes the `plugins` system. See `Readme` for information on how to convert from `plugins` to using `AuthForm` and `applyAxiosInterceptors`
 - Removes the `EnvVar` system. It wasn't being used except for Auth Profiles which are now handled outside of BadMagic through interceptors
+- Removes support for `sticky` attribute on a route
 
 ## v0.0.28